==> 61-topic-diet.Rmd <==
# 肠道菌群与饮食 {#diet}

## 核心文章列表



```{r diet-article-datatable, fig.cap="饮食相关研究的核心论文"}
topic_header <- topic[["diet"]]
topic_caption <- paste0(topic_header,"研究核心论文")

# 文章列表
df <- topic_articles_core[["diet"]]
df <- biblio_df(df)
produce_topic_article_datatable(df, caption = topic_caption, filename = topic_caption)
```

首先奉上“饮食”相关研究的关键论文清单，如果有《热心肠日报》解读的话，也一并附上（图 \@ref(fig:diet-article-datatable)）。
这里包括了 `r nrow(df)` 篇论文，包括 `r nrow(dplyr::filter(df,DT=="REVIEW"))` 篇高水平综述，其中有中文短科普的就有 `r nrow(dplyr::filter(df,daily_cn != ""))` 篇。 
如果你是一个初学者，可按此图索骥，快速把握肠道菌群与饮食相关的研究内容。



## 领域内专家  

图 \@ref(fig:diet-article-top-author) 中列出了本领域知名专家，其中就包括我们前面(\@ref(top-scientist))介绍过的两位最出众的科学家：
KNIGHT, ROB 和 GORDON, JEFFERY I.。应该说，后者在饮食研究领域的贡献更为突出，因为你将会看到历史脉络中的绝大多数文章都是由他们课题组完成的。


```{r diet-article-top-author, fig.cap="饮食相关研究的顶尖学者"}
topAU <- authorProdOverTime(df, k=10, graph = TRUE)
```


从作者合作网络中（图 \@ref(fig:diet-article-author-collaboration-network)），可以发现 Rob Knight 和 Jeffrey I. Gordon 是来自于同一个研究群体，这一群体还有 3 个重要人物，分别是 Bernard Henrissat，Dan Knights 和 Federico E. Rey。

除此之外，我们还可以发现 5 个其他的研究群体。

- Hazen, Stanley L., Wang, Zeneng, Tang, W. H. Wilson, Wu, Yuping 和 Lusis, Aldons J.

- Cryan, John F., Dinan, Timothy G. 等人

- Cani, Partrice D., Delzenne, Nathalie M. 等人

- Flint, Harry J., Duncan, Sylvia H. 等人

- Ventura, Marco 等人



```{r diet-article-author-collaboration-network, fig.cap="饮食相关研究的主要作者合作网络。显示发文量最多的 30 位研究者之间的合作关系。"}
## Author collaboration
author_network(df,nNode = 30)
```


## 机构合作网络

从发文量看机构排行，前30家机构中领衔的圣路易斯华盛顿大学（WASHINGTON UNIV）、哥本哈根大学（UNIV COPENHAGEN）和科罗拉多大学（UNIV COLORADO）等（图 \@ref(fig:diet-univeristy-top30)）。

```{r diet-univeristy-top30, fig.cap="饮食相关研究发文量最高的30家机构"}
tableTag_barplot(df,Tag = "AU_UN_NR") +
  labs(title = "")
```
观察机构间的合作网络，可以发现科罗拉多大学和天主教鲁汶大学（CATHOLIC UNIV LOUVAIN）处于网络中非常核心的地（图 \@ref(fig:diet-article-university-collaboration-network)）。哥本哈根大学与哥德堡大学（UNIV GOTHENBURG）之间的合作强度比较大。

```{r diet-article-university-collaboration-network, fig.cap="饮食相关研究机构合作网络\n这个网络中的节点取自本领域发文量最大的30家机构，机构间的连线表明二者存在合作发表论文的情况。连线的线越粗，说明二者间的论文发表数量越多。点的大小越大，说明其在合作网络中处于更重要的节点位置。为了更好的显示效果，本图经过适当简化，去掉了孤立的点和较弱的连接。"}
## 机构合作网络（显示前 15 家）
university_network(df, nNode = 30, edge_weight_cutoff = 3)
```

## 研究热点

从这些研究使用的关键词，可以瞥见研究的热点。最受关注的30个关键词中，最值得关注的是（短）链脂肪酸（Chain Fatty-acids），
这一个关键词的文章数量仅次于肠道菌群的两种不同表达方法（Gut microbiota 和 intestinal microbiota），排在了多样性（Diversity）的前面。

而肥胖（Obesity）作为与饮食关系最为密切的一种常见病，排在了第6位。其它一些疾病相关的关键词，包括炎症、IBD（Inflammatory-bowel-disease）、
胰岛素抗性（Insulin resistance）、IBS（Irritable bowel syndrome）、克罗恩病（Crohns Disease）等
（图 \@ref(fig:diet-id-top30)）。

```{r diet-id-top30, fig.cap="以文章数量计，最受关注的关键词（Top30）"}
tableTag_barplot(df, Tag = "ID")
```

这些关键词组成了如下所示的一个共现网络（图 \@ref(fig:diet-id-top30)）。
不出意外，肥胖和（短）链脂肪酸在网络中占据最突出的地位。
可以发现，蛋白偶联受体（Protein Coupled Receptor）和（短）链脂肪酸之间有着非常强的联系，
意味着前者在后者的功能中具有重要的作用。

```{r diet-id-network, fig.cap="关键词共现网络包括了相关研究中最长出现的50个关键词。\n为突出网络主体，一些孤立的节点和较弱的连接被隐去。"}
keyword_network(df,nNode = 50,edge_weight_cutoff = 10)
```


```{r eval=FALSE}

KeywordGrowth(df,top = 30) %>% pivot_longer(-Year) %>%
  ggplot(aes(Year,name,size=value,color=value)) +
  geom_point(alpha=1/3) +
  scale_size(range = c(1,10))
  
```


```{r diet-article-conceptual-structure-map, fig.cap="饮食相关研究共词分析", eval=FALSE}
# 共词分析
safe_conceptualStructure <- possibly(conceptualStructure,NULL)
CS <- safe_conceptualStructure(df, field = "AB", method = "MDS",clust=5, minDegree = round(nrow(df)/5), stemming = FALSE, labelsize = 10, documents = 30,graph = FALSE)
CS$graph_terms
```


```{r diet-article-thematicmap, fig.cap="饮食相关研究的研究主题地图", eval=FALSE}
## Thematic Map
safe_thematicMap <- purrr::possibly(thematicMap, NULL)
Map <- safe_thematicMap(df, field="ID",n=250,minfreq = 5,stemming = FALSE,size = 0.5, n.labels = 3, repel = TRUE)
Map$map
```

## 肠道菌群与饮食研究的历史脉络


```{r}
histPlot <- function (histResults, n = 20, size = 5, labelsize = 5, verbose = TRUE) 
{
  colorlist <- c(brewer.pal(9, "Set1")[-6], brewer.pal(8, 
    "Set2")[-7], brewer.pal(12, "Paired")[-11], brewer.pal(12, 
    "Set3")[-c(2, 8, 12)])
  if (isTRUE(size)) {
    size <- 5
  }
  LCS <- colSums(histResults$NetMatrix)
  NET <- histResults$NetMatrix
  s = sort(LCS, decreasing = TRUE)[min(n, length(LCS))]
  ind = which(LCS >= s)
  NET = NET[ind, ind]
  LCS = LCS[ind]
  bsk.network <- graph_from_adjacency_matrix(NET, mode = c("directed"), 
    weighted = NULL)
  R <- strsplit(names(V(bsk.network)), ",")
  RR <- lapply(R, function(l) {
    l = l[1:2]
    l = paste(l[1], l[2], sep = ",")
  })
  V(bsk.network)$id <- unlist(RR)
  deg <- LCS
  V(bsk.network)$size <- size
  Years <- as.numeric(unlist(str_extract_all(unlist(RR), "[[:digit:]]{4}$")))
  V(bsk.network)$years <- Years
  bsk.network <- simplify(bsk.network, remove.multiple = T, 
    remove.loops = T)
  E(bsk.network)$color <- "slategray1"
  bsk.network <- bibliometrix:::delete.isolates(bsk.network)
  dg <- decompose.graph(bsk.network)
  layout_m <- create_layout(bsk.network, layout = "nicely")
  layout_m$cluster <- 0
  rr <- 0
  for (k in 1:length(dg)) {
    bsk <- dg[[k]]
    a <- ifelse(layout_m$name %in% V(bsk)$name, k, 0)
    layout_m$cluster <- layout_m$cluster + a
    Min <- min(layout_m$y[layout_m$cluster == k]) - 1
    layout_m$y[layout_m$cluster == k] <- layout_m$y[layout_m$cluster == 
      k] + (rr - Min)
    rr <- max(layout_m$y[layout_m$cluster == k])
  }
  bsk <- bsk.network
  wp <- membership(cluster_infomap(bsk, modularity = FALSE))
  layout_m$x <- layout_m$years
  g <- ggraph(layout_m) + geom_edge_arc(width = 1, strength = 0.05, 
    check_overlap = T, edge_alpha = 0.5, color = "grey", 
    arrow = grid::arrow(angle = 10, unit(0.15, "inches"))) + 
    geom_node_text(aes(label = V(bsk)$id), size = labelsize, 
      repel = TRUE, color = colorlist[wp], alpha = 0.8) + 
    geom_node_point(size = V(bsk)$size, color = "royalblue4", 
      alpha = 0.15) + scale_color_brewer() + scale_x_continuous(labels = as.character(seq(min(Years), 
    max(Years))), breaks = seq(min(Years), max(Years))) + 
    theme_minimal() + theme(legend.position = "none", panel.background = element_rect(fill = "gray97", 
    color = "grey97"), axis.line.y = element_blank(), axis.text.y = element_blank(), 
    axis.ticks.y = element_blank(), axis.title.y = element_blank(), 
    axis.title.x = element_blank(), axis.line.x = element_line(size = 3, 
      colour = "grey80"), panel.grid.minor.y = element_blank(), 
    panel.grid.major.y = element_blank(), panel.grid.minor.x = element_blank(), 
    axis.text.x = element_text(face = "bold")) + labs(title = "Historical Direct Citation Network")
  plot(g)
  if (isTRUE(verbose)) {
    cat("\n Legend\n\n")
    label <- data.frame(Label = names(V(bsk.network)), stringsAsFactors = FALSE)
    Data <- histResults$histData
    Data <- left_join(label, Data, by = c(Label = "Paper"))
    print(Data[, -2])
  }
  return(bsk.network)
}
```


接下来，我们通过观察这些文章的**历史引证网络**，来理清相关研究的脉络（图 \@ref(fig:diet-article-histnet)）。

```{r diet-article-histnet, fig.cap="饮食相关文献的核心引证网络"}
safe_histNetwork <- purrr::possibly(histNetwork, NULL)
hist_result <- safe_histNetwork(filter(df,DT=="ARTICLE"),verbose = FALSE)
hisgraph <- histPlot(hist_result,verbose = F,n=20,size = 8,labelsize = 3)
```


### 菌群被发现是饮食与肥胖之间重要的一环

从图 \@ref(fig:diet-article-histnet) 中可以看到，最早的一篇研究可以追溯到 2004 年[@backhedGutMicrobiotaEnvironmental2004]。
这篇发表在 PNAS 上的研究论文使用无菌小鼠做了饲喂实验，结果发现正常小鼠体内的肠道菌群能够促进无菌小鼠对糖类的吸收，并促进脂肪的累积。
而如果没有肠道菌群的作用，无菌小鼠中的脂肪累积就会显著下降。不仅如此，该研究还鉴定出小鼠中的 Fiaf 基因在微生物参与的脂肪累积中发挥重要作用，从而说明了肠道菌群是一个影响饮食能量吸收和储存的重要因素。

同样是  Bäckhed F，同样是在 PNAS 上面，2007 年他们又发表了一篇研究论文，阐明了无菌小鼠之所以不会肥胖的分子机制[@backhedMechanismsUnderlyingResistance2007]。
这项研究针对前一项研究中一个有意思的现象展开，那就是无菌小鼠在面对高脂、高糖的西式饮食时，较有菌小鼠更加不容易出现脂肪累积。
原因是什么呢？一种途径可能在于肠道细胞中通常被抑制的 Fiaf 基因的上调，另一种途径在于骨骼肌和肝脏中磷酸化的AMPK水平的上升。
由于菌群是 Fiaf 基因表达的一个抑制因素，所以当菌群不存在时，Fiaf 基因就会上调，导致下游脂肪酸代谢基因的失调，不利于脂肪累积。
而AMPK下游相关的脂肪酸代谢基因同样也是如此。所以，两种相互独立的信号通路参与了饮食因素在小鼠肥胖中的作用，而根源都与肠道菌群息息相关。

同年发表的另一篇文章，将肠道菌群之所以能够促进肥胖的关键诱发因素给鉴定了出来，那就是细菌分泌的内毒素，具体而言是细菌的脂多糖（LPS）[@caniMetabolicEndotoxemiaInitiates2007]。
CANI PD 等的研究发现，细菌脂多糖能够进入人体血液并作用于 CD14 细胞，从而引起代谢性的炎症，最终导致全身性的胰岛素抵抗，提高小鼠肥胖和糖尿病风险。
而不管是降低 LPS 浓度还是突变 CD14，都能够抑制小鼠糖尿病和肥胖症的发作。


#### 不论是长期还是短期，饮食都是肠道菌群最重要的影响因素之一

在另一篇论文中，研究人员分析了饮食饮食在婴儿肠道菌群发育中可能的作用[@palmerDevelopmentHumanInfant2007]。
这项研究一共纳入了 14 个健康的婴儿，包括一对同卵双胞胎。因为早期研究方法的局限性，作者用了非常多的方法来跟踪婴儿肠道菌群随时间的变化过程，
其中包括基因芯片、qPCR、16S rRNA基因克隆和测序等。最终分析了 548 个样本，包括婴儿的粪便、母亲的阴道拭子和粪便、父亲和兄弟姐妹的粪便，以及母乳等等。
并初步分析了疾病、药物、饮食变化和旅行等关键事件对菌群的影响。

TURNBAUGH PJ 2009年的另一个研究，选取成年女性同卵及异卵双胞胎（同一对双胞胎胖瘦一致），分析其与其母亲的粪便菌群，研究宿主基因型、环境暴露及宿主肥胖如何影响肠道菌群[@turnbaughCoreGutMicrobiome2009]。一共分析了154个个体，结果显示家庭成员之间既共享部分肠道菌群，也存在群落组成的差异，差异程度在同卵双胞胎与异卵双胞胎中相似。同时发现，在基因水平而菌群谱系水平，所有个体拥有一个核心菌群。


```{r filippoImpactDietShaping2010-fig1,fig.cap="非洲乡村的风貌和居民饮食[@filippoImpactDietShaping2010]"}
knitr::include_graphics("images/filippoImpactDietShaping2010_fig1.png")
```

2010年DE FLILIPPO 发表在 PNAS 上的那篇论文，通过比较欧洲和非洲农村地区儿童的肠道菌群和饮食差异，发现饮食是塑造肠道菌群的关键因素[@filippoImpactDietShaping2010]。在非洲农村（图 \@ref(fig:filippoImpactDietShaping2010-fig1)），儿童饮食中含有更多的植物多糖，相应地造成肠道菌群中有更加丰富的拟杆菌门细菌，更加少的厚壁菌门细菌，以及非常丰富的普雷沃氏菌和 *Xylanibacter*。后者被认为在纤维素和木聚糖的水解中发挥重要作用，在欧洲儿童肠道菌群中没有发现。

WALKER AW 2011年的论文研究了10周的时间跨度内，不同类型饮食对志愿者肠道菌群的调节作用[@walkerDominantDietresponsiveGroups2011]。研究除了对照组，还有高抗性淀粉饮食、非淀粉多糖饮食和低糖饮食等3种。2名志愿者的结果显示抗性淀粉饮食中 60% 的抗性淀粉仍然没有被消耗，这使得肠道菌群产生了明显的变化，包括瘤胃球菌和 *Oscillibacter* 等厚壁菌门细菌的显著增加。
这表明食物中不可被人类利用的成分会对肠道菌群的结构产生显著影响。


沿着这条线走下去，是 TURNBAUGH PJ 在2008/2009年发表的三篇重要研究。在 2008 年发表的 Cell Host and Microbe 杂志的研究中报道了肥胖与小鼠远端肠道菌群改变相关[@turnbaughDietInducedObesityLinked2008]，主要表现在两个方面：一是发现西式饮食能够促进厚壁菌门柔膜菌纲的激增，而限制饮食控制肥胖后这种现象消失；二是发现将这个胖鼠的菌群移植到瘦的无菌小鼠中后，会大大促进瘦鼠脂肪的堆积。

在 2009 年发表在 Science Translational Medicine 杂志上的研究，则通过将人的肠道菌群移植到无菌小鼠中，研究“人鼠合一”的肠道菌群如何被饮食调控[@turnbaughEffectDietHuman2009]。
结果发现，首先人的肠道菌群是可以在小鼠内定殖的，且定殖后的菌群特征与人类的相似。其次，当从低脂、富含膳食纤维的饮食转向高脂高糖的西式饮食后，
可在一天内改变肠道菌群的结构，以及菌群的代谢、基因表达等。本研究同时也发现，与定殖的历史相比，饮食是改变菌群更重要的因素。


而在 2011 年，PNAS 上的另一项研究通过观察婴儿肠道菌群的演替，最终明确婴儿微生物组的改变与生活事件（包括饮食、健康状态等）是密切相关的[@koenigSuccessionMicrobialConsortia2011]。
辅食的添加是促进婴儿肠道菌群转向成人的重要因素。

进一步地，在成人中也观测到饮食对肠道菌群的显著调节作用[@wuLinkingLongtermDietary2011]。食用**高脂/低膳食纤维**或者**低脂/高膳食纤维**的饮食，24小时内就可以改变肠道菌群。前者引起拟杆菌的上升，后者引起普雷沃氏菌的上升。鉴于这两种菌是肠型的主要定义类群，因此长期的饮食作用下，会改变一个人的肠型。这一现象不仅出现在人身上，其它哺乳动物中也是如此[@mueggeDietDrivesConvergence2011]。紧接着，在2014年，DAVID LA 在 Nature 杂志发文阐释了**动物来源食物**和**植物来源食物**短期内对肠道菌群的影响[@davidDietRapidlyReproducibly2014]，他们同时发现，动物来源的食物通过*Bilophila wadsworthia* 与炎症性肠病关联了起来。


2012 年发表在 Nature 的两篇论文关注了更多人群中肠道菌群与饮食和健康之间的关系[@claessonGutMicrobiotaComposition2012, @yatsunenkoHumanGutMicrobiome2012]。通过比较婴儿、儿童和成人的肠道菌群差异，明确了肠道菌群随年龄的变化趋势[@yatsunenkoHumanGutMicrobiome2012]。此外，也发现美国与其它国家居民的肠道菌群存在差异[@yatsunenkoHumanGutMicrobiome2012]。虽然地域也是菌群差异的重要因素，但是最直接的原因仍然被锚定在了饮食上。例如：与地理位置相比，饮食是老年人肠道菌群差异的主要决定因素[@claessonGutMicrobiotaComposition2012]。

#### 饮食演变为一种重要的干预手段

SOKOL H 2008年发表在 PNAS 上的研究，主要是鉴定了克罗恩病患者肠道内一种抗炎共生细菌 *Faecalibacterium Prausnitzii*[@sokolFaecalibacteriumPrausnitziiAntiinflammatory2008]。
*F. Prausnitzii*是厚壁菌门的一员，在CD复发患者中的比例显著降低。而该菌存在时，其分泌的代谢产物能够抑制肠道细胞的炎症反应。
这一研究虽然并不直接与饮食相关，却是饮食干预治疗 CD 的重要基础。


饮食对肠道菌群结构性的改变只是表象，与之相比，更重要的是肠道菌群对食物中不同营养成分代谢能力的差异。这不仅影响人的消化能力，还会改变一些疾病的患病风险。2011年 WANG ZN 发表在 Nature 杂志上的论文，就发现菌群的代谢产物氧化三甲胺（TMAO）是心血管疾病重要的风险因素[@wangGutFloraMetabolism2011]。TMAO由肠道菌群代谢胆碱后生成，吸收进入血液后可以促进动脉粥样硬化。在易患动脉粥样硬化的小鼠中抑制肠道菌群，可抑制饮食中胆碱增多带来的动脉粥样硬化。

随后，在2013年 Nature Medicine 上，KOETH RA 发表论文表示红肉中的肉碱是肠道菌群代谢产生 TMAO 的可用来源之一[@koethIntestinalMicrobiotaMetabolism2013]，从而解开了红肉与动脉粥样硬化之间的内在关联。

RIDAURA VK 2013年发表在 Science 上面的论文，用人粪菌移植的方法，描述了通过调节菌群改善肥胖的可能性[@ridauraGutMicrobiotaTwins2013a]：[移植胖人粪菌的小鼠，长得更胖](https://www.mr-gut.cn/papers/read/1050193804)。

自此，饮食开始成为一种重要的干预手段，被广泛应用于疾病治疗、健康保健中。

### 养出好菌群

历史研究脉络展示了2014年前的20篇关键论文，而饮食与肠道菌群的故事当然还在继续。
通过前面的这些研究，我们意识到肠道菌群具有人类没有的代谢能力，主要是大多数的植物多糖都不会被人类消化，所以会进入结肠成为肠道菌群的潜在食物来源。
细菌不同菌株具有分解不同底物的多样性和诱导能力，整个菌群具有数千个参与碳水化合物分解代谢的基因[@farrellImportanceFeedingYour2019]。

值得注意的是，肠道菌群中的一些成员，不仅可以利用食物中的聚糖成分，在膳食纤维不足的情况下，还会去分解肠道内的粘膜聚糖，从而导致结肠粘膜屏障的侵蚀。
所以养出好菌群非常重要！你得给它吃菜，如果没有菜它只能去啃你的肠子[@desaiDietaryFiberDeprivedGut2016]。
这可能是饮食诱发肠道菌群失调后引起肥胖、炎症性肠病等肠道疾病的重要途径之一。

反过来，如果我们能够弄清楚不同的肠道菌群成员对营养的不同偏好，那么调节肠道菌群将变成一种操作性很强的工作。
例如，2018 年 Nature 上面发表的一篇论文，
使用 [改造菌株+配套营养支持=可精细调控的菌株定殖](https://www.mr-gut.cn/papers/read/1065476652)。
他们向拟杆菌株中转化可利用紫菜聚糖的罕见基因位点，使之能利用紫菜聚糖作为碳源；
通过喂食紫菜聚糖，在小鼠肠道中形成一种独有的代谢生态位，以供这种改造的外源菌株独享，使该菌株可以在含有不同肠道菌群的小鼠中成功植入；
调节紫菜聚糖的喂食剂量，可精细调控该菌株在小鼠肠道内的丰度，且该效果可逆；
这种靶向的膳食营养支持，足以使改造的外源菌株与已在小鼠结肠隐窝内定殖的同基因型菌株竞争，并取代后者。


未来，如果我们弄清楚了菌群中不同组分在健康或者疾病中的作用，那么靶向菌群的“精准营养”或许可以成为一种非常有效的治疗手段。
