==> 65-topic-disease-cancer.Rmd <==
## 癌症  {#cancer}

### 癌症相关研究的核心论文



```{r crc-articles, fig.cap="本表格仅列出了相关研究中关键论文"}
crc_result <- disease_dt_output("cancer")
crc <- crc_result$data
crc_result$widget
```

与肿瘤相关的研究一共有 `r nrow(crc)` 篇关键论文发表。
图 \@ref(fig:crc-articles) 列出了这些论文。借助这些文章，可以快速掌握这一领域研究的主要成果。





```{r}
four_dimension_barplot_caption <- paste0("肿瘤",
                                         "相关研究的简要信息。",
                                         "（A）发表文章最多的10个国家；",
                                         "（B）发表文章最多的10家机构；",
                                         "（C）发表文章最多的10位科学家；",
                                         "（D）发表文章最多的10本杂志。")
```


```{r crc-four-dimension-barplot, fig.cap=four_dimension_barplot_caption}
four_dimension_barplot(crc)
```
简单统计一下，可以发现相关研究的主体来源于美国、中国、英国和意大利等国（图 \@ref(fig:crc-four-dimension-barplot)A）。

十大顶尖的研究机构中，哈佛大学及其分支占了其中的一半以上（图 \@ref(fig:crc-four-dimension-barplot)B），
包括哈佛大学、丹那-法博癌症研究所、哈佛医学院、麻省总医院、哈佛大学公共卫生学院和布莱根妇女医院等六家，
其余机构也多位于美国，包括加州大学圣地亚哥分校、国立癌症研究所NCI等。
此外，法国国家健康医学研究院（INSERM）位列第六，实力不俗。
中国的上海交通大学屈居第十，是唯一上榜的国内研究机构。

在十大科学家的名单中，我们很高兴的发现了一个中国人的名字 YU JUN（图 \@ref(fig:crc-four-dimension-barplot)C），
不出意外的话，这应该是来自于中国香港的于君教授。
于教授在肠道菌群与癌症的研究中蜚声海内外，她能进入 Top10 可谓是实至名归。

在十大来源期刊中，GUT、Science、Cell Host & Microbe 杂志拔得前三（图 \@ref(fig:crc-four-dimension-barplot)D）。


### 癌症相关研究的合作网络

进一步地，我们看看相关机构和科学家的合作网络（图 \@ref(fig:crc-university-net)）。

美国的哈佛系、欧洲的法国班底，是世界上两大合作团体。

```{r crc-university-net, fig.cap="癌症研究的机构合作网络"}
university_network(crc,edge_weight_cutoff = 3,nNode = 30,delete_isolate=FALSE) 
```


我们取前 30 位发表文章数量最多的作者，分析他们的合作网络（图 \@ref(fig:crc-author-net)）。
几乎绝大多数的研究者，都存在着彼此之间的联系，又组成了三个联系非常紧密的学术群体。
而中国相关的于君教授，则相对比较独立的开展了相关研究，其主要的合作者只有两位老师。

```{r crc-author-net, fig.cap="癌症研究的作者合作网络"}
author_network(crc,edge_weight_cutoff = 1,nNode = 30,delete_isolate=FALSE) 
```

### 肠癌相关研究的历史脉络

#### 具核梭杆菌——肠癌的重要诱因

通过历史引证网络梳理相关研究的脉络。在图 \@ref(fig:crc-histplot) 中，列出了 20 篇最关键的研究论文。
这 20 篇论文明显分成了两个部分，在图中分别位列上半块区域和下半块区域。

```{r crc-histplot, fig.cap="癌症相关研究的历史脉络"}
hist_graph <- histPlot(histNetwork(crc %>% filter(DT == "ARTICLE"), verbose = FALSE), verbose = FALSE)
```

这些论文的具体信息可以参阅图 \@ref(fig:crc-histplot-articles)。

最早的一项重要研究发表在 Nature Medcine 杂志上，发现了一种可以产毒素的脆弱拟杆菌可以诱导小鼠肿瘤的发生，
并指出肿瘤发生的过程依赖于 Th17 细胞介导的免疫反应通路[@wuHumanColonicCommensal2009]。

2011年发表在 PLoS One 杂志上的一篇论文，则首次用 16S rRNA 基因测序的方芳，
比较分析了大肠癌患者和非大肠癌患者粪便菌群的差异[@sobhaniMicrobialDysbiosisColorectal2011]。
研究一共分析了 12 个粪便样本，其中一半是大肠癌患者。
PCA 分析结果很清晰的显示，大肠癌患者具有与正常人不同的粪便菌群。


2011年发表在 PLoS One 杂志上的另一篇论文，则首次用 16S rRNA 基因测序的方法，
比较分析了人大肠癌组织和正常组织中的微生物群落分布情况[@marchesiHumanColorectalCancer2011]。
这一项研究详细报道了两个微环境下的群落结构差异，并发现结肠细菌科在肿瘤组织中富集，
而肠杆菌科的细菌则在肿瘤组织中下降。

2012年发表在 Science 上的研究，主要关注了炎症反应在肠道菌群诱导癌症发生过程中的关键作用。
《热心肠日报》解读了 [这篇文章](https://www.mr-gut.cn/papers/read/1070902056)，
简而言之就是：炎症毁菌，菌增炎症，终而癌也！


2012年背靠背发表在 Genome Research 上的两篇论文，明确了致癌的一个关键物种。
其中的一项研究发现了梭杆菌在肠癌组织中富集，
为揭示诱导癌症发生的肠道共生菌指出了一条关键的方向[@kosticGenomicAnalysisIdentifies2012]。
而另一项研究则直接指出具核梭杆菌（*Fusobacterium nucleatum*）的感染，除能够引发牙周炎和阑尾炎之外，
还与肠癌发生存在关联[@castellarinFusobacteriumNucleatumInfection2012]。

2012年发表在 ISME J 杂志上的一篇研究，有上海交通大学的赵立平团队完成[@wangStructuralSegregationGut2012a]。
在这项研究中，一共分析了 46 个大肠癌患者和 56 个健康志愿者的肠道菌群，
进一步确认了二者之间存在的肠道菌群差异，并鉴定出 11 个在大肠癌患者中富集的 OTU，
以及 5 个在大肠癌患者中下降的 OTU。


```{r crc-histplot-articles, fig.cap="癌症相关研究的关键节点文献"}
cap <- "癌症相关研究的关键节点文献"
vip <- extract_from_hist_graph(crc, hist_graph)
produce_topic_article_datatable(vip, cap, cap)
```

2013年发表在 JNCI 杂志上的研究同样是用 16S rRNA 基因测序的方法比较了大肠癌患者与健康人的肠道菌群差异。
研究纳入了 47 例患者和 94 例对照，最终指出梭杆菌（*Fusobacterium*）和
卟啉单胞菌（*Porphyromonas*）可能在致癌过程中的作用[@ahnHumanGutMicrobiome2013]。

具核梭杆菌与肠癌之间的联系，在 2013 - 2014 年还有另外的多篇研究发表，其中 Cell Host & Microbe 两篇，
PLoS One 一篇，PNAS 一篇，Cancer Research 一篇，等等。不仅明确了具核梭杆菌的致癌作用、机制，
还尝试利用此类标记为进行早期的诊断等，在此不一一赘述。
总之，具核梭杆菌就这样成了臭名昭著的致癌细菌，再无翻身之处。

#### 肠道菌群影响癌症治疗的效果

在历史引证网络的上半部分，是一个有所不同的研究领域（图 \@ref(fig:crc-histplot)）。
这几篇研究主要关注肠道菌群对癌症治疗药物的作用，以及对癌症治疗效果的影响。

先是发现抗癌药的是否有效，肠道菌群说了算。进而认识到，在肠道菌群的助力下，一些抗癌药的作用可被显著加强。
并给出了使用阿克曼氏菌等益生菌能够促进癌症治疗效果的例子。

这些论文均有《热心肠日报》的解读，在此一笔带过。

需要说明的是，此处的癌症其实不仅仅是肠道癌症，这些研究属于肠道菌与药物相互作用的范畴。


总之，肠道菌群与癌症的关系可能存在两个方面，一是肠道菌群在消化道癌症发病过程中起到的作用，
二是肠道菌群对癌症治疗可能存在的影响。


### 癌症相关的重要研究主题

```{r}
crc_keyword_net_cap <- paste0("癌症相关的研究主题。",
                              "在这个共现网络中，显示最多不超过 50 个最常见的节点，",
                              "且为突出网络主题，去掉了类似于 Gut Microbiota 的一般词汇",
                              "以及共现次数小于 3 次的连接")
```


在癌症相关研究中出现最多的关键词组成了共现网络，从中可以看出相关研究的一些主要方向。
在这其中，短链脂肪酸、炎症是最突出的研究热点，它们与膳食纤维、调节性T细胞、癌症发生等
方面存在千丝万缕的联系（图 \@ref(fig:crc-keyword-net)）。
而具核梭杆菌成为联系二者的重要桥梁之一。

```{r crc-keyword-net, fig.cap=crc_keyword_net_cap}
keyword_network(crc,edge_weight_cutoff = 3,nNode = 50,delete_isolate=FALSE,
                remove_keyword = search_keyword_regex) %>%
  visOptions(manipulation = TRUE)

```
